+++ src/idunno.Bluesky/PostBuilder.cs
@@ -3,6 +3,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 
+using idunno.AtProto;
 using idunno.AtProto.Repo;
 
 using idunno.Bluesky.Actor;