+++ src/idunno.Bluesky/Record/Profile.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
 
+using idunno.AtProto;
 using idunno.AtProto.Labels;
 using idunno.AtProto.Repo;