+++ src/idunno.Bluesky/RichText/HashTag.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using idunno.AtProto;
+
 namespace idunno.Bluesky.RichText
 {
     /// <summary>