+++ test/idunno.Bluesky.Test/RichTextTests.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Diagnostics.CodeAnalysis;
-
 namespace idunno.Bluesky.Test
 {
     [ExcludeFromCodeCoverage]